11.26

+++ generate_multi.py
@@ -58,7 +58,7 @@ if __name__ == '__main__':
         
     # 加载模型和tokenizer
     tokenizer = AutoTokenizer.from_pretrained(model_dir)
-    model = AutoModelForCausalLM.from_pretrained(model_dir).to(device)
+    model = AutoModelForCausalLM.from_pretrained(model_dir, dtype=torch.float16).to(device)
     if tokenizer.pad_token_id is None:
         tokenizer.pad_token = tokenizer.eos_token
